remove su/sudo from user runit service
Both su and sudo use fork instead of execve to run a
command for a user. As a result, runit monitors su/sudo instead of the
user-specific runsvdir process and it cannot send the proper signals to
reload or terminate runsvdir.

This commit replaces su with chpst. chpst doesn't set up the user
environment, so we do that manually with printenv.