depend on the new chruby cookbook
The only icky consequence is that we still need to use our own chruby.sh
source script, so right now the chruby.sh is converged back and forth
for each run. An option to disable the chruby cookbook from spitting out
the chruby.sh file might be the Right Move(TM).